+++ src/main/java/br/caixa/gov/credito/apisimulador/service/dto/ParcelaResponseDTO.java
@@ -1,5 +1,11 @@
 package br.caixa.gov.credito.apisimulador.service.dto;
 
-    public record ParcelaResponseDTO(Integer numero, Double valorAmortizacao, Double valorJuros, Double valorPrestacao ) {
+import java.math.BigDecimal;
+
+public record ParcelaResponseDTO(
+    Integer numero, 
+    BigDecimal valorAmortizacao, 
+    BigDecimal valorJuros, 
+    BigDecimal valorPrestacao ) {
     }
